requirements.txt
Flask == 2.0.0
Flask-SQLAlchemy==2.5.1
gunicorn == 20.1.0
pytesseract == 0.3.7
opencv-contrib-python ==4.5.2.52
tesserocr == 2.5.1
Pillow == 7.0.0
numpy == 1.18.1
numpydoc == 0.9.2
opencv-python-headless==4.5.2.52
